requirements.txt
line-bot-sdk
flask
configparser
langchain
PyPDF2
qdrant_client
langchain_openai
pypdf
openai